+++ api/user.js
@@ -1,0 +1,5 @@
+export default {
+    getMe(app) {
+      return app.$axios.$post('/api/me')
+    },
+  }
