--- store/index.js
@@ -1,15 +1,0 @@
-import Vuex from 'vuex'
-import auth from './modules/auth'
-import user from './modules/user'
-
-const store = () =>
-  new Vuex.Store({
-    modules: {
-      auth,
-      user
-    },
-    strict: false,
-    plugins: []
-  })
-
-export default store
